backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
yfinance==0.2.35
python-docx>=1.1.0
anthropic>=0.40.0